@@ -2,8 +2,8 @@ package com.nbk.dao.converters;
 
 import com.nbk.dao.domain.account.TransactionTypeEnum;
 
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 import java.util.stream.Stream;
 
 /**
